fix: refactor of classes & added constructors for component in newGameCallback to fix errors

+++ client/src/scenes/game/Game.cpp
@@ -7,12 +7,13 @@
 
 #include "Scenes.hpp"
 #include "Gameplay.hpp"
+#include "Components.hpp"
 
 void newGameCallback(Registry &r)
 {
     r.clear_enities();
     entity_t player = r.spawn_entity();
-    r.add_component(player, Components::Entity(0, 0, 50, 50));
+    r.add_component(player, Components::Entity(0, 0, static_cast<std::size_t>(50), static_cast<std::size_t>(50)));
     r.add_component(player, Components::Player(100, "Player", 0));
     r.add_component(player, Components::Drawable(Textures::PLAYER_ID));
     r.add_component(player, Components::Movable(1));
